+++ package.json
@@ -39,13 +39,13 @@
     "prepublishOnly": "npm test && npm run build"
   },
   "dependencies": {
-    "fp-ts": "^1.19.0",
-    "io-ts": "^1.8.0"
+    "fp-ts": "^2.0.5",
+    "io-ts": "^2.0.1"
   },
   "devDependencies": {
     "@contactlab/tslint-config": "1.3.0",
     "@types/jest": "24.0.18",
-    "@types/node": "^12.0.0",
+    "@types/node": "^12.7.3",
     "husky": "3.0.4",
     "jest": "24.9.0",
     "jest-fetch-mock": "2.1.2",
@@ -56,22 +56,22 @@
     "ts-node": "8.3.0",
     "tslint": "5.19.0",
     "tslint-config-prettier": "1.18.0",
-    "typescript": "3.5.3"
+    "typescript": "3.6.2"
   },
   "jest": {
     "automock": false,
     "bail": true,
     "coveragePathIgnorePatterns": [
-      "<rootDir>/test/_helpers/",
+      "<rootDir>/src/test/_helpers/",
       "<rootDir>/node_modules/"
     ],
     "coverageReporters": [
       "text"
     ],
     "roots": [
-      "<rootDir>/test/"
+      "<rootDir>/src/test/"
     ],
-    "testRegex": "(\\.|/)spec\\.(js|ts)$",
+    "testRegex": "(\\.|/)spec\\.ts$",
     "moduleFileExtensions": [
       "js",
       "json",
