+++ tsconfig-compile.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "include": ["./src/**/*"]
+  "exclude": ["./src/test"]
 }
